+++ db/migrate/20260511140100_create_installations.rb
@@ -2,9 +2,9 @@ class CreateInstallations < ActiveRecord::Migration[8.1]
   def change
     create_table :installations do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :github_installation_id, null: false
+      t.bigint :github_installation_id, null: false
       t.string :account_login, null: false
-      t.integer :account_id, null: false
+      t.bigint :account_id, null: false
       t.string :account_type, null: false
       t.datetime :installed_at, null: false
       t.datetime :removed_at
